+++ tests/e2e/ec_shardcache_eval_test.go
@@ -123,7 +123,7 @@ func TestE2E_ECShardCacheEval(t *testing.T) {
 		accessKey, secretKey = bootstrapAdminViaUDSAny(t, dataDirs[:1], 60*time.Second)
 
 		for i := 1; i < numNodes; i++ {
-			require.NoError(t, writeNodeJoinPending(dataDirs[i], raftAddr(0)))
+			require.NoError(t, writeNodeJoinPending(dataDirs[i], dataDirs[0], raftAddr(0)))
 			procs[i] = startNode(i)
 			time.Sleep(150 * time.Millisecond)
 		}
@@ -402,7 +402,7 @@ func TestE2E_ECShardCacheActive(t *testing.T) {
 		accessKey, secretKey = bootstrapAdminViaUDSAny(t, dataDirs[:1], 60*time.Second)
 
 		for i := 1; i < numNodes; i++ {
-			require.NoError(t, writeNodeJoinPending(dataDirs[i], raftAddr(0)))
+			require.NoError(t, writeNodeJoinPending(dataDirs[i], dataDirs[0], raftAddr(0)))
 			procs[i] = startNode(i)
 			time.Sleep(150 * time.Millisecond)
 		}